Prepare for graceful expansion
In preparation for adding modes which handle brownouts,
I separated the update() which is called from loop(), and
the definition of the mode's specific update code into
modeUpdate.  This allows modeUpdate to be written without
having to call the parent update() method (which will
likely get forgotten).

Also added modeUpdate() periodic calling, and changed
start() and stop(0 to match this pattern.