Working - with known issues
Issue 1. Rounding.  If the decimals round up to a new whole, it won't
register in the whole part.

Issue 2. Fixed buffer size - we should have a sensive default, like 30
but allow the user to specify another amount in the constructor.

Issue 3. There's no concenient function to indicate that a whole string
of numbers has finished playing... busy() will intermittently return a
wrong value when between numbers