move MutilaNeoPixel to sepate lib
Turns out that adding this was kindof a pain in the butt
with respect to maintaing backwards compatibility with
old projects, as the NeoPixel thing added a dep to the
Adafruit_NeoPixel library.  This means that it's really
not possible to cleanly keep old projects building without
modification in the Makefile or elsewhere in the case of
Arduino GUI builds... so I've put it in a separate library.