Replace DB with Serial.print
It doesn't make sense in the IDE to use DB and DBLN, as there is no way
to add or remove -DDEBUG from it. This isn't a problem when there is
just an .ino file in an example, because we can put the #define DEBUG
directly in the sketch, but when we have an example with other source
files, this is impossible.

DB is reserved just for awesome makefile users.